chore: 修改缓存方式

+++ src/router/index.js
@@ -109,6 +109,16 @@ router.beforeEach(async (to, form, next) => {
     const token = localStorage.getItem('token');
     const corpId = localStorage.getItem('corpId');
     const src = window.location.pathname;
+    // if (!token && !options.code && options.appid) {
+    if (!token && options.appid && options.appid !== corpId && !options.code) {
+      const sourceId = options.channel || '';
+      window.location.href = `https://open.weixin.qq.com/connect/oauth2/authorize?appid=${
+        options.appid
+      }&redirect_uri=${
+        encodeURIComponent(`${Config.redirect_uri}${src}?channel=${sourceId}&appid=${options.appid}&batchNo=${options.batchNo || ''}`)
+      }&response_type=code&scope=snsapi_userinfo&state=${sourceId}#wechat_redirect`;
+      return;
+    }
     if (options.appid !== corpId) {
       localStorage.clear();
     }
@@ -138,16 +148,6 @@ router.beforeEach(async (to, form, next) => {
         if (options.channel) {
           localStorage.setItem('channel', options.channel);
         }
-      }
-      // if (!token && !options.code && options.appid) {
-      if (!token && options.appid && options.appid !== corpId && !options.code) {
-        const sourceId = options.channel || '';
-        window.location.href = `https://open.weixin.qq.com/connect/oauth2/authorize?appid=${
-          options.appid
-        }&redirect_uri=${
-          encodeURIComponent(`${Config.redirect_uri}${src}?channel=${sourceId}&appid=${options.appid}&batchNo=${options.batchNo || ''}`)
-        }&response_type=code&scope=snsapi_userinfo&state=${sourceId}#wechat_redirect`;
-        return;
       }
     }
   }
